refactor(tests): split large test files into smaller focused modules to comply with LOC limits
