+++ tests/vibe3/clients/test_github_client_prs.py
@@ -317,235 +317,3 @@ def test_create_pr_repairs_empty_body_after_creation(
 def review_mixin():
     """Create ReviewMixin instance."""
     return ReviewMixin()
-
-
-def test_get_pr_diff_file_limit_error(review_mixin):
-    """Test that PR diff with >300 files raises UserError."""
-    with patch("subprocess.run") as mock_run:
-        # Mock subprocess error with file limit message
-        mock_run.side_effect = subprocess.CalledProcessError(
-            returncode=1,
-            cmd=["gh", "pr", "diff", "200"],
-            stderr=(
-                "could not find pull request diff: HTTP 406: Sorry,"
-                " the diff exceeded the maximum number of files (300)."
-                " Consider using 'List pull requests files' API"
-                " or locally cloning the repository instead."
-            ),
-        )
-
-        with pytest.raises(UserError) as exc_info:
-            review_mixin.get_pr_diff(200)
-
-        error_msg = str(exc_info.value)
-        assert "too many files" in error_msg
-        assert "GitHub limit: 300" in error_msg
-        assert "#200" in error_msg
-        assert "vibe inspect branch" in error_msg
-
-
-def test_get_pr_diff_other_error(review_mixin):
-    """Test that other PR diff errors raise GitHubError."""
-    with patch("subprocess.run") as mock_run:
-        mock_run.side_effect = subprocess.CalledProcessError(
-            returncode=1,
-            cmd=["gh", "pr", "diff", "200"],
-            stderr="Network error",
-        )
-
-        with pytest.raises(GitHubError) as exc_info:
-            review_mixin.get_pr_diff(200)
-
-        assert exc_info.value.status_code == 1
-        assert "Network error" in exc_info.value.message
-
-
-def test_get_pr_files_file_limit_error(review_mixin):
-    """Test that PR files with >300 files raises UserError."""
-    with patch("subprocess.run") as mock_run:
-        # Mock subprocess error with file limit message
-        mock_run.side_effect = subprocess.CalledProcessError(
-            returncode=1,
-            cmd=["gh", "pr", "diff", "200", "--name-only"],
-            stderr=(
-                "could not find pull request diff: HTTP 406: Sorry,"
-                " the diff exceeded the maximum number of files (300)."
-            ),
-        )
-
-        with pytest.raises(UserError) as exc_info:
-            review_mixin.get_pr_files(200)
-
-        error_msg = str(exc_info.value)
-        assert "too many files" in error_msg
-        assert "GitHub limit: 300" in error_msg
-        assert "#200" in error_msg
-
-
-def test_get_pr_files_success(review_mixin):
-    """Test successful get_pr_files."""
-    with patch("subprocess.run") as mock_run:
-        mock_result = MagicMock()
-        mock_result.stdout = "file1.py\nfile2.py\nfile3.py\n"
-        mock_result.returncode = 0
-        mock_run.return_value = mock_result
-
-        result = review_mixin.get_pr_files(42)
-
-        assert result == ["file1.py", "file2.py", "file3.py"]
-
-
-def test_get_pr_files_other_error(review_mixin):
-    """Test that other PR files errors raise GitHubError."""
-    with patch("subprocess.run") as mock_run:
-        mock_run.side_effect = subprocess.CalledProcessError(
-            returncode=1,
-            cmd=["gh", "pr", "diff", "200", "--name-only"],
-            stderr="Authentication failed",
-        )
-
-        with pytest.raises(GitHubError) as exc_info:
-            review_mixin.get_pr_files(200)
-
-        assert exc_info.value.status_code == 1
-        assert "Authentication failed" in exc_info.value.message
-
-
-def test_error_message_suggests_alternatives(review_mixin):
-    """Test that error message suggests alternative approaches."""
-    with patch("subprocess.run") as mock_run:
-        mock_run.side_effect = subprocess.CalledProcessError(
-            returncode=1,
-            cmd=["gh", "pr", "diff", "200"],
-            stderr="diff exceeded the maximum number of files (300)",
-        )
-
-        with pytest.raises(UserError) as exc_info:
-            review_mixin.get_pr_diff(200)
-
-        error_msg = str(exc_info.value)
-        # Should suggest alternatives
-        assert "Alternatives:" in error_msg
-        assert "vibe inspect branch" in error_msg
-        assert "pull/200/files" in error_msg
-
-
-def _client() -> GitHubClient:
-    return GitHubClient()
-
-
-# -- list_issues_with_assignees --
-
-
-def test_list_issues_with_assignees_success():
-    client = _client()
-    payload = [{"number": 1, "title": "T", "labels": [], "assignees": [], "url": "u"}]
-    mock_result = MagicMock(returncode=0, stdout=json.dumps(payload))
-    with patch("subprocess.run", return_value=mock_result):
-        result = client.list_issues_with_assignees(limit=10)
-    assert len(result) == 1
-    assert result[0]["number"] == 1
-
-
-def test_list_issues_with_assignees_failure_returns_empty():
-    client = _client()
-    mock_result = MagicMock(returncode=1, stderr="auth error", stdout="")
-    with patch("subprocess.run", return_value=mock_result):
-        result = client.list_issues_with_assignees()
-    assert result == []
-
-
-def test_list_issues_with_assignees_passes_repo():
-    client = _client()
-    mock_result = MagicMock(returncode=0, stdout="[]")
-    with patch("subprocess.run", return_value=mock_result) as mock_run:
-        client.list_issues_with_assignees(repo="owner/repo")
-    cmd = mock_run.call_args[0][0]
-    assert "--repo" in cmd
-    assert "owner/repo" in cmd
-
-
-# -- close_issue --
-
-
-def test_close_issue_success():
-    client = _client()
-    mock_result = MagicMock(returncode=0, stdout="", stderr="")
-    with patch("subprocess.run", return_value=mock_result):
-        assert client.close_issue(42) is True
-
-
-def test_close_issue_with_comment():
-    client = _client()
-    mock_result = MagicMock(returncode=0, stdout="", stderr="")
-    with patch("subprocess.run", return_value=mock_result) as mock_run:
-        client.close_issue(42, comment="closing because done")
-    cmd = mock_run.call_args[0][0]
-    assert "--comment" in cmd
-
-
-def test_close_issue_failure_returns_false():
-    client = _client()
-    mock_result = MagicMock(returncode=1, stderr="not found", stdout="")
-    with patch("subprocess.run", return_value=mock_result):
-        assert client.close_issue(999) is False
-
-
-# -- add_comment --
-
-
-def test_add_comment_success():
-    client = _client()
-    mock_result = MagicMock(returncode=0, stdout="", stderr="")
-    with patch("subprocess.run", return_value=mock_result):
-        assert client.add_comment(42, "hello") is True
-
-
-def test_add_comment_failure_returns_false():
-    client = _client()
-    mock_result = MagicMock(returncode=1, stderr="error", stdout="")
-    with patch("subprocess.run", return_value=mock_result):
-        assert client.add_comment(42, "hello") is False
-
-
-def test_add_comment_passes_repo():
-    client = _client()
-    mock_result = MagicMock(returncode=0, stdout="", stderr="")
-    with patch("subprocess.run", return_value=mock_result) as mock_run:
-        client.add_comment(42, "hi", repo="owner/repo")
-    cmd = mock_run.call_args[0][0]
-    assert "--repo" in cmd
-
-
-# -- get_pr_for_issue --
-
-
-def test_get_pr_for_issue_found():
-    client = _client()
-    payload = [
-        {
-            "number": 10,
-            "closingIssuesReferences": [{"number": 42}],
-        }
-    ]
-    mock_result = MagicMock(returncode=0, stdout=json.dumps(payload))
-    with patch("subprocess.run", return_value=mock_result):
-        pr = client.get_pr_for_issue(42)
-    assert pr == 10
-
-
-def test_get_pr_for_issue_not_found():
-    client = _client()
-    payload = [{"number": 10, "closingIssuesReferences": [{"number": 99}]}]
-    mock_result = MagicMock(returncode=0, stdout=json.dumps(payload))
-    with patch("subprocess.run", return_value=mock_result):
-        pr = client.get_pr_for_issue(42)
-    assert pr is None
-
-
-def test_get_pr_for_issue_gh_failure():
-    client = _client()
-    mock_result = MagicMock(returncode=1, stderr="error", stdout="")
-    with patch("subprocess.run", return_value=mock_result):
-        pr = client.get_pr_for_issue(42)
-    assert pr is None
