+++ tests/vibe3/orchestra/test_governance_service.py
@@ -309,259 +309,3 @@ class TestGovernanceService:
 
         await service.on_tick()
         assert called["count"] == 1
-
-    def test_build_governance_plan_can_disable_skill_content(self, tmp_path):
-        """Governance prompt template should allow omitting skill body."""
-        prompts_path = tmp_path / "prompts.yaml"
-        prompts_path.write_text(
-            "orchestra:\n"
-            "  governance:\n"
-            "    plan: |\n"
-            "      Skill={skill_name}\n"
-            "      Content={skill_content}\n"
-        )
-        snapshot = OrchestraSnapshot(
-            timestamp=0.0,
-            server_running=True,
-            active_issues=(),
-            active_flows=0,
-            active_worktrees=0,
-            circuit_breaker_state="closed",
-            circuit_breaker_failures=0,
-        )
-        service = GovernanceService(
-            config=OrchestraConfig(
-                governance=GovernanceConfig(
-                    dry_run=True,
-                    prompt_template="orchestra.governance.plan",
-                    include_skill_content=False,
-                )
-            ),
-            status_service=MockStatusService(snapshot),
-            dispatcher=_make_dispatcher(),
-            prompts_path=prompts_path,
-        )
-
-        plan = service._build_governance_plan(snapshot)
-        assert "Skill=vibe-orchestra" in plan
-        assert "Content=" in plan
-        assert "# Vibe Orchestra" not in plan
-
-    def test_build_material_source_summary_reports_configured_sources(
-        self, tmp_path, monkeypatch
-    ):
-        """Dry run metadata should point to the configured prompt and skill sources."""
-        prompts_path = tmp_path / "prompts.yaml"
-        prompts_path.write_text(
-            "orchestra:\n" "  governance:\n" "    plan: |\n" "      Prompt body\n"
-        )
-        skill_path = tmp_path / "skills" / "demo" / "SKILL.md"
-        skill_path.parent.mkdir(parents=True)
-        skill_path.write_text("# Demo Skill\n", encoding="utf-8")
-        monkeypatch.chdir(tmp_path)
-
-        service = GovernanceService(
-            config=OrchestraConfig(
-                governance=GovernanceConfig(
-                    skill="demo",
-                    prompt_template="orchestra.governance.plan",
-                )
-            ),
-            status_service=MockStatusService(),
-            dispatcher=_make_dispatcher(),
-            prompts_path=prompts_path,
-        )
-
-        summary = service._build_material_source_summary()
-
-        assert summary["prompt_template_key"] == "orchestra.governance.plan"
-        assert summary["prompt_template_file"] == str(prompts_path)
-        assert summary["skill_name"] == "demo"
-        assert summary["skill_file"] == str(skill_path)
-
-    @pytest.mark.asyncio
-    async def test_run_governance_honors_governance_dry_run(self, tmp_path):
-        """Dry run stops before dispatch and persists a preview plan under temp/."""
-        snapshot = OrchestraSnapshot(
-            timestamp=0.0,
-            server_running=True,
-            active_issues=(),
-            active_flows=0,
-            active_worktrees=0,
-            circuit_breaker_state="closed",
-            circuit_breaker_failures=0,
-        )
-        dispatcher = _make_dispatcher()
-        dispatcher.repo_path = tmp_path
-        service = GovernanceService(
-            config=OrchestraConfig(governance=GovernanceConfig(dry_run=True)),
-            status_service=MockStatusService(snapshot),
-            dispatcher=dispatcher,
-        )
-
-        await service._run_governance()
-        dispatcher.run_governance_command.assert_not_called()
-
-        dry_run_files = sorted((tmp_path / "temp").glob("governance_dry_run_*.md"))
-        assert len(dry_run_files) == 1
-        assert "# Orchestra Governance Scan" in dry_run_files[0].read_text()
-
-
-class TestGovernanceRecipeDrivenRendering:
-    """Assert governance uses PromptAssembler (recipe-driven) rendering."""
-
-    def test_render_uses_assembler_not_direct_format(self, tmp_path):
-        """GovernanceService._render_governance_plan should use PromptAssembler."""
-
-        prompts_path = tmp_path / "prompts.yaml"
-        prompts_path.write_text(
-            "orchestra:\n"
-            "  governance:\n"
-            "    plan: |\n"
-            "      Skill={skill_name}\n"
-            "      Status={server_status}\n"
-            "      Content={skill_content}\n"
-            "      Running={running_issue_count}\n"
-            "      Suggested={suggested_issue_count}\n"
-            "      Flows={active_flows}\n"
-            "      CB={circuit_breaker_state}\n"
-            "      CBF={circuit_breaker_failures}\n"
-            "      Running Issues:\n"
-            "      {running_issue_details}\n"
-            "      Suggested Issues:\n"
-            "      {suggested_issue_details}\n"
-            "      Active={active_count}\n"
-            "      Worktrees={active_worktrees}\n"
-            "      Issues:\n"
-            "      {issue_list}\n"
-            "      {truncated_note}\n"
-        )
-        service = GovernanceService(
-            config=OrchestraConfig(
-                governance=GovernanceConfig(
-                    prompt_template="orchestra.governance.plan",
-                    include_skill_content=False,
-                )
-            ),
-            status_service=MockStatusService(),
-            dispatcher=_make_dispatcher(),
-            prompts_path=prompts_path,
-        )
-        # Verify _render_governance_plan uses assembler internally
-        # by checking the result is a PromptRenderResult or string but via assembler
-        snapshot = OrchestraSnapshot(
-            timestamp=0.0,
-            server_running=True,
-            active_issues=(),
-            active_flows=2,
-            active_worktrees=1,
-            circuit_breaker_state="closed",
-            circuit_breaker_failures=0,
-        )
-        context = service._build_prompt_context(snapshot)
-        result_text = service._render_governance_plan(context)
-        assert "Skill=vibe-orchestra" in result_text
-        assert "Status=running" in result_text
-        assert "Flows=2" in result_text
-
-    def test_dry_run_exposes_render_result_provenance(self, tmp_path):
-        """Dry-run result should carry PromptRenderResult with provenance."""
-        prompts_path = tmp_path / "prompts.yaml"
-        prompts_path.write_text(
-            "orchestra:\n"
-            "  governance:\n"
-            "    plan: |\n"
-            "      Skill={skill_name}\n"
-            "      Status={server_status}\n"
-            "      Content={skill_content}\n"
-            "      Running={running_issue_count}\n"
-            "      Suggested={suggested_issue_count}\n"
-            "      Flows={active_flows}\n"
-            "      CB={circuit_breaker_state}\n"
-            "      CBF={circuit_breaker_failures}\n"
-            "      Running Issues:\n"
-            "      {running_issue_details}\n"
-            "      Suggested Issues:\n"
-            "      {suggested_issue_details}\n"
-            "      Active={active_count}\n"
-            "      Worktrees={active_worktrees}\n"
-            "      Issues:\n"
-            "      {issue_list}\n"
-            "      {truncated_note}\n"
-        )
-        service = GovernanceService(
-            config=OrchestraConfig(
-                governance=GovernanceConfig(
-                    prompt_template="orchestra.governance.plan",
-                    include_skill_content=False,
-                )
-            ),
-            status_service=MockStatusService(),
-            dispatcher=_make_dispatcher(),
-            prompts_path=prompts_path,
-        )
-        snapshot = OrchestraSnapshot(
-            timestamp=0.0,
-            server_running=True,
-            active_issues=(),
-            active_flows=0,
-            active_worktrees=0,
-            circuit_breaker_state="closed",
-            circuit_breaker_failures=0,
-        )
-        context = service._build_prompt_context(snapshot)
-        # render should return a string (render result text), internally using assembler
-        plan_text = service._render_governance_plan(context)
-        assert isinstance(plan_text, str)
-        # The service should expose the last render result for dry-run logging
-        render_result = service.last_render_result
-        assert render_result is not None
-        assert render_result.recipe_key == "orchestra.governance.plan"
-        prov_vars = {p.variable for p in render_result.provenance}
-        assert "skill_name" in prov_vars
-        assert "server_status" in prov_vars
-
-    def test_governance_recipe_uses_skill_source_for_skill_content(self, tmp_path):
-        """skill_content variable should come from SKILL source in the recipe."""
-        from vibe3.prompts.models import VariableSourceKind
-
-        prompts_path = tmp_path / "prompts.yaml"
-        prompts_path.write_text(
-            "orchestra:\n"
-            "  governance:\n"
-            "    plan: |\n"
-            "      Skill={skill_name}\n"
-            "      Status={server_status}\n"
-            "      Content={skill_content}\n"
-            "      Running={running_issue_count}\n"
-            "      Suggested={suggested_issue_count}\n"
-            "      Flows={active_flows}\n"
-            "      CB={circuit_breaker_state}\n"
-            "      CBF={circuit_breaker_failures}\n"
-            "      Running Issues:\n"
-            "      {running_issue_details}\n"
-            "      Suggested Issues:\n"
-            "      {suggested_issue_details}\n"
-            "      Active={active_count}\n"
-            "      Worktrees={active_worktrees}\n"
-            "      Issues:\n"
-            "      {issue_list}\n"
-            "      {truncated_note}\n"
-        )
-        service = GovernanceService(
-            config=OrchestraConfig(
-                governance=GovernanceConfig(
-                    prompt_template="orchestra.governance.plan",
-                    include_skill_content=True,
-                    skill="vibe-orchestra",
-                )
-            ),
-            status_service=MockStatusService(),
-            dispatcher=_make_dispatcher(),
-            prompts_path=prompts_path,
-        )
-        recipe = service._build_governance_recipe()
-        skill_src = recipe.variables.get("skill_content")
-        assert skill_src is not None
-        assert skill_src.kind == VariableSourceKind.SKILL
-        assert skill_src.skill == "vibe-orchestra"
